Refactor calculations controller. Move methods into Helper file and some more minor refactoring
